requirements.txt
torch~=2.6.0
opencv-python~=4.10.0.84
pillow~=11.1.0
torchvision~=0.21.0
pandas~=2.2.3
numpy~=1.26.4
matplotlib~=3.10.0
scikit-learn~=1.6.1